Extract shared database test-harness loaders into a package
The .env reading, migration-role connection resolver, and per-test
schema-name generator lived inside the hub-api git harness, where no
other test project could reach them. Move them into a private
@intx/test-harness workspace package so any test suite that needs a
real database connection resolves it the same way, through the repo's
existing node_modules workspace links rather than a cross-project path
import.
