+++ package.json
@@ -6,7 +6,8 @@
   "workspaces": [
     "packages/*",
     "apps/*",
-    "examples/*"
+    "examples/*",
+    "tests/lib"
   ],
   "catalog": {
     "@types/semver": "^7.7.1",
@@ -21,6 +22,7 @@
   },
   "devDependencies": {
     "@eslint/js": "9.34.0",
+    "@intx/test-harness": "workspace:*",
     "@types/bun": "1.3.9",
     "arktype": "catalog:",
     "drizzle-kit": "^0.31.9",
