+++ tsconfig.json
@@ -48,6 +48,7 @@
     { "path": "examples/ring-demo" },
     { "path": "tests/hub-agent" },
     { "path": "tests/hub-api" },
+    { "path": "tests/lib" },
     { "path": "tests/tool-packaging" },
     { "path": "tests/workflow-deploy" },
     { "path": "bin" }
